requirements.txt
botorch==0.6.0
dill==0.3.5.1
geopy==2.2.0
gpytorch==1.6.0
gym==0.21.0
gymnasium==0.28.1
isaacgym.egg==info
isaacgymenvs.egg==info
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.1
opencv_python==4.7.0.72
pandas==1.4.1
PyAutoGUI==0.9.53
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
sb3_contrib==2.0.0a8
scipy==1.7.3
setuptools==45.2.0
sympy==1.11.1
torch==2.0.0
wandb==0.12.21
